+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - d10p defense die.
+### Fixed
+- Error messages when adding innate weapons.
 
 ## [0.3.8] - 2022-12-28
 ### Fixed